task-manager/python/requirements.txt
fastapi
uvicorn
python-jose
passlib
pydantic
httpx
supabase